No real changes
just pokes and prods at the wall run lerp class which I need to remember to rename. I'm gonna need to make an isolated test for this, but I need to make a test lerp scenario without a player controller to see if the lerp function can move an object at all. If so, the player controller will need its own dedicated and custom lerp function.